repo cleanup before aoc2022

+++ src/lib.rs
@@ -6,7 +6,7 @@ pub mod nom;
 pub fn chardigit(c: u8) -> u8 {
     const ZERO: u8 = b'0';
     assert!(c.is_ascii_digit());
-    (c as u8) - ZERO
+    c - ZERO
 }
 
 pub fn expect_empty_line<S: AsRef<str>>(s: S) -> anyhow::Result<()> {
@@ -53,6 +53,7 @@ macro_rules! derive_FromStr_for_FromIterator {
     ($t:ty, $v:ty) => {
         impl ::core::str::FromStr for $t {
             type Err = <$v as ::core::str::FromStr>::Err;
+
             #[inline]
             fn from_str(
                 s: &str,
@@ -68,6 +69,7 @@ macro_rules! derive_FromStr_for_bytes_TryFrom_collect {
     ($t:ty, $v:ty) => {
         impl ::core::str::FromStr for $t {
             type Err = <$v as ::core::convert::TryFrom<u8>>::Error;
+
             #[inline]
             fn from_str(
                 s: &str,
@@ -95,6 +97,7 @@ macro_rules! derive_FromStr_for_nom {
     ($t:ty, $f:path) => {
         impl ::core::str::FromStr for $t {
             type Err = ::nom::error::Error<String>;
+
             #[inline]
             fn from_str(s: &str) -> ::core::result::Result<Self, Self::Err> {
                 match ::nom::Finish::finish(($f)(s)) {
